+++ playerbot/strategy/triggers/RpgTriggers.cpp
@@ -117,7 +117,7 @@ bool RpgBuyTrigger::IsActive()
     if (!guidP.HasNpcFlag(UNIT_NPC_FLAG_VENDOR))
         return false;
 
-    if (AI_VALUE(uint8, "durability") < 50)
+    if (AI_VALUE(float, "durability") < 50)
         return false;
 
     if (AI_VALUE(bool, "can sell")) //Need better condition.
@@ -159,7 +159,7 @@ bool RpgAHBuyTrigger::IsActive()
     if (!guidP.HasNpcFlag(UNIT_NPC_FLAG_AUCTIONEER))
         return false;
 
-    if (AI_VALUE(uint8, "durability") < 50)
+    if (AI_VALUE(float, "durability") < 50)
         return false;
 
     if (AI_VALUE(bool, "can ah sell")) //Need better condition.
